Revert incorrect RCL_3 drop: Revision: 201033 Kit: 201035

+++ secureswitools/swianalysistoolkit/source/dumpswicertstoretool/swicertstoremetadata.cpp
@@ -143,7 +143,7 @@ const string SwiCertStoreMetaData ::ExtractCapabilities(char* aCapability, int a
    				capability.append("Unknown"+unknownIndex);
    				unknownIndex++;
    				}
-  			cap += pow((double)2,j);
+  			cap += pow(2,j);
    			}
    		else
    			{
